Refactor and clean up upload files
